Removed fullscreen toggle and other things
removed tutorial.cs because useless, added an fps counter because why not, GlobalTimer is now a long, so sadly no more short waits for an overflow

+++ Assets/scripts/Global/Timer.cs
@@ -4,7 +4,7 @@ namespace Global
 {
     public class Timer : MonoBehaviour
     {
-        public short GlobalTimer;
+        public long GlobalTimer;
         // Start is called before the first frame update
         void Start()
         {
